feat(diet): 영양 요약 토글을 오른쪽 끝으로 옮기고 주간 그래프를 홈과 같은 꺾은선으로
